Fix output interleaving with spinner
Unfortunately when we use the spinner library, it only works when you work
on 1 line that is the last line of the output which is _the_ spinner.

If you print anything after that, it will be overwritten by spinner, hence
showing weird output like in #61, then re-printing the spinner twice.

Fixing handling of "FYI, running..." output to be consistent with other FYIs.

Fixes #61.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>