add pause period to readiness check
Currently we probably are hammering the api during the 4 minute period
non-stop as quick as the GetService returns. This might cause users to
potentially hit a rate limit error (causing run-button to fail itself).

So adding a 2 second time period to cool things down. Since it corresponds
to max 120 calls over 240 secs it should be ok for now.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>