provide env ordering via "order" key (#209)
Introducing an optional "order" key to the env object.

The sorting takes a look at if "order" field is specified or not. If not,
the key is prompted last. If specified, it sorts them by the value of the
"order" field.

The order is still non-deterministic if env objects don't specify an
"order" value. (i.e. this patch still doesn't preserve map order).

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>